requirements.txt
Django>=1.9
django-appconf>=1.0.2
django-setuptest>=0.2.1
requests>=2.11.0
requests-oauthlib>=0.6.2
django-user-accounts>=2.0.3